+++ DSTools/HighLevelDirServices/HighLevelDirServicesMini.h
@@ -1,10 +1,28 @@
 /*
+ * Copyright (c) 2001-2005 Apple Computer, Inc. All rights reserved.
+ *
+ * @APPLE_LICENSE_HEADER_START@
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
+ * 
+ * The Original Code and all software distributed under the License are
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
+ * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
+ * 
+ * @APPLE_LICENSE_HEADER_END@
+ */
+
+/*
  *  HighLevelDirServices.h
- *  DirTesting
- *
- *  Created by fhill on Mon Nov 05 2001.
- *  Copyright (c) 2001 Apple Computer, Inc. All rights reserved.
- *
  */
 
 #ifndef _HIGHLEVELDIRSERVICES_H_
@@ -41,7 +59,7 @@ tDirStatus				HLDSGetAttributeValuesFromRecordsByAttributeValues( const tDirRefe
 tDirStatus				HLDSSetAttributeValue( const tDirReference inDirRef, tRecordReference inRecordRef, const char* inAttributeName,
 							char createIfNecessary, const char* inAttributeValue );
 tDirStatus				HLDSSetAttributeValues( const tDirReference inDirRef, tRecordReference inRecordRef, const char* inAttributeName,
-							char createIfNecessary, const char** inAttributeValues, unsigned long inNumValues );
+							char createIfNecessary, const char** inAttributeValues, UInt32 inNumValues );
 tDirStatus				HLDSSetAttributeCFValues( const tDirReference inDirRef, tRecordReference inRecordRef, const char* inAttributeName,
 							char createIfNecessary, CFArrayRef inAttributeValues );
 tDirStatus				HLDSSetBinaryAttributeValue( const tDirReference inDirRef, tRecordReference inRecordRef, const char* inAttributeName,
