+++ libsecurity_asn1/lib/SecNssCoder.cpp
@@ -67,7 +67,7 @@ SecNssCoder::~SecNssCoder()
 
 PRErrorCode	SecNssCoder::decode(
 	const void				*src,		// BER-encoded source
-	unsigned long			len,
+	size_t				len,
 	const SecAsn1Template 	*templ,	
 	void					*dest)
 {
